+++ pexdoc/exdoc.py
@@ -1,6 +1,6 @@
 # -*- coding: utf-8 -*-
 # exdoc.py
-# Copyright (c) 2013-2017 Pablo Acosta-Serafini
+# Copyright (c) 2013-2018 Pablo Acosta-Serafini
 # See LICENSE for details
 # pylint: disable=C0111,C0411,C0413,E0012,E0611,E1101,E1103,F0401,R0201,R0903
 # pylint: disable=R0913,R0914,W0105,W0122,W0212,W0611,W0613
@@ -239,6 +239,7 @@ class ExDocCxt(object):
         # Restore exception handler (if any) active when context was activated
         if self._existing_exhobj is not None:
             pexdoc.exh.set_exh_obj(self._existing_exhobj)
+        return True
 
 
 ###